requirements.txt
appdirs==1.4.3
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
black==18.9b0
bleach==3.1.0
bravado==10.2.2
bravado-core==5.10.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.3.2
defusedxml==0.5.0
docopt==0.6.2
entrypoints==0.3
factory-boy==2.11.1
Faker==1.0.2
Flask==1.0.2
genson==1.0.2
googlemaps==3.0.2
grip==4.5.2
idna==2.8
inflection==0.3.1
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jsonref==0.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Markdown==2.6.11
MarkupSafe==1.1.0
mistune==0.8.4
monotonic==1.5
more-itertools==5.0.0
msgpack-python==0.5.6
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
pandocfilters==1.4.2
parso==0.3.2
path-and-address==2.0.1
pickleshare==0.7.5
pluggy==0.8.1
prettytable==0.7.2
prometheus-client==0.5.0
prompt-toolkit==2.0.8
py==1.7.0
Pygments==2.3.1
pytest==4.1.1
pytest-factoryboy==2.0.2
python-dateutil==2.7.5
python-jsonschema-objects==0.3.5
pytz==2018.9
pywinpty==0.5.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
rfc3987==1.3.8
ruamel.yaml==0.15.87
Send2Trash==1.5.0
simplejson==3.16.0
six==1.12.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.3
terminado==0.8.1
testpath==0.4.2
text-unidecode==1.2
toml==0.10.0
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webcolors==1.8.1
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
